start resolving comments
