+++ antarest/core/core_blueprint.py
@@ -14,13 +14,9 @@
 from fastapi import APIRouter
 
 from antarest.core.config import Config
-from antarest.core.requests import UserHasNotPermissionError
 from antarest.core.serde import AntaresBaseModel
 from antarest.core.utils.web import APITag
 from antarest.core.version_info import VersionInfoDTO, get_commit_id, get_dependencies
-from antarest.login.utils import require_current_user
-from antarest.matrixstore.model import MatrixMismatchDTO
-from antarest.matrixstore.service import ISimpleMatrixService
 
 
 class StatusDTO(AntaresBaseModel):
@@ -31,7 +27,7 @@ class DesktopModeDTO(AntaresBaseModel):
     desktop_mode: bool
 
 
-def create_utils_routes(config: Config, matrix_service: ISimpleMatrixService) -> APIRouter:
+def create_utils_routes(config: Config) -> APIRouter:
     """
     Utility endpoints
 
@@ -78,22 +74,5 @@ def create_utils_routes(config: Config, matrix_service: ISimpleMatrixService) ->
         """
 
         return DesktopModeDTO(desktop_mode=config.desktop_mode)
-
-    @bp.post(
-        "/private/resolve-matrix-store",
-        summary="Synchronize Database with filesystem for the matrix-store. To be used if an issue occurred",
-    )
-    def synchronize_matrix_store(dry_run: bool) -> dict[str, MatrixMismatchDTO]:
-        """
-        To be used by the admin only.
-        If `dry_run` is True, only returns the list of mismatches. Else, also performs the 2 following operations:
-        - Deletes lines from `Matrix` table in DB for matrices that do not exist on the filesystem.
-        - Insert lines inside `Matrix` table in DB for matrices that exist on the filesystem but not in the DB.
-        """
-        user = require_current_user()
-        if not user.is_site_admin():
-            raise UserHasNotPermissionError()
-
-        return matrix_service.synchronize_matrix_store(dry_run=dry_run)
 
     return bp
