+++ tests/core/test_utils_bp.py
@@ -34,7 +34,7 @@ def test_version() -> None:
     mock_storage_service.study_service.path_resources = Path("/")
 
     app = FastAPI(title=__name__)
-    app.include_router(create_utils_routes(Config(), Mock()))
+    app.include_router(create_utils_routes(Config()))
     client = TestClient(app)
 
     path = "/version"
@@ -46,7 +46,7 @@ def test_version() -> None:
 
 def test_server_health() -> None:
     app = FastAPI(title=__name__)
-    app.include_router(create_utils_routes(Config(), Mock()))
+    app.include_router(create_utils_routes(Config()))
     client = TestClient(app)
     result = client.get("/health")
     assert result.json() == {"status": "available"}
